+++ pakyow-assets/spec/unit/types/js_spec.rb
@@ -3,12 +3,12 @@ require "pakyow/application/behavior/assets/types/js"
 RSpec.describe Pakyow::Application::Behavior::Assets::Types::Js do
   let :config do
     app_class = Class.new(Pakyow::Application) do
-      include Pakyow::Application::Config::Assets
-
       configurable :presenter do
         setting :path, ""
       end
     end
+
+    app_class.include_framework :assets
 
     app_class.config.assets.tap do |config|
       config.source_maps = false