+++ pakyow-support/lib/pakyow/support/configurable/config.rb
@@ -3,9 +3,12 @@
 require "concurrent/array"
 require "concurrent/hash"
 
+require "pakyow/support/class_state"
 require "pakyow/support/deep_dup"
 require "pakyow/support/deep_freeze"
 require "pakyow/support/deprecatable"
+require "pakyow/support/makeable"
+require "pakyow/support/object_name"
 
 require "pakyow/support/configurable/setting"
 
@@ -22,91 +25,142 @@ module Pakyow
 
         extend Pakyow::Support::Deprecatable
 
-        # @api private
-        attr_reader :__settings, :__defaults, :__groups
+        extend Pakyow::Support::ClassState
+        class_state :__settings, default: Concurrent::Hash.new, inheritable: true
+        class_state :__defaults, default: Concurrent::Hash.new, inheritable: true
+        class_state :__groups, default: Concurrent::Hash.new, inheritable: true
 
-        def initialize(configurable, name: nil, path: [])
-          @configurable, @__name, @__path = configurable, name, path
+        include Pakyow::Support::Hookable
+        include Pakyow::Support::Makeable
 
-          @__settings = Concurrent::Hash.new
-          @__defaults = Concurrent::Hash.new
-          @__groups = Concurrent::Hash.new
-          @__deprecations = Concurrent::Array.new
+        after "make" do
+          __settings.each_value do |setting|
+            setting.update_configurable(__configurable)
+          end
         end
 
-        def initialize_copy(_)
-          @__defaults = @__defaults.deep_dup
-          @__settings = @__settings.deep_dup
-          @__groups = @__groups.deep_dup
-          @__deprecations = @__deprecations.deep_dup
+        class << self
+          attr_reader :object_name
 
-          @__settings.each_pair do |key, _|
-            define_setting_methods(key)
+          # @api private
+          def __configurable
+            @configurable
           end
 
-          @__groups.each_pair do |key, _|
-            define_group_methods(key)
-          end
-
-          @__deprecations.each do |deprecation|
-            define_deprecation deprecation[:target], deprecation[:solution]
-          end
-
-          super
-        end
-
-        # Define setting `name` with `default`. If a block is given, the default value will be built
-        # eagerly the first time the setting is accessed.
-        #
-        def setting(name, default = default_omitted = true, &block)
-          tap do
+          # Define setting `name` with `default`. If a block is given, the default value is built
+          # and cached the first time the setting is accessed.
+          #
+          def setting(name, default = default_omitted = true, &block)
             name = name.to_sym
 
             if default_omitted
               default = nil
             end
 
-            unless @__settings.include?(name)
+            unless __settings.include?(name)
               define_setting_methods(name)
             end
 
-            @__settings[name] = Setting.new(default: default, configurable: @configurable, &block)
+            __settings[name] = Setting.new(
+              name: name,
+              default: default.deep_dup,
+              configurable: __configurable,
+              &block
+            )
+
+            self
+          end
+
+          # Define defaults for one or more environments.
+          #
+          def defaults(*environments, &block)
+            environments.each do |environment|
+              (__defaults[environment] ||= Concurrent::Array.new) << block
+            end
+          end
+
+          # Define a nested configurable.
+          #
+          def configurable(group, &block)
+            group = group.to_sym
+
+            config = Config.make(
+              ObjectName.build(
+                *object_name.parts.reject { |part| part == :config }, group
+              ),
+              context: self,
+              configurable: __configurable
+            )
+
+            config.class_eval(&block)
+
+            unless __groups.include?(group)
+              define_group_methods(group)
+            end
+
+            __groups[group] = config
+          end
+
+          # Configure default values for `environment`.
+          #
+          def configure_defaults!(environment)
+            __defaults[environment.to_s.to_sym].to_a.each do |defaults|
+              instance_eval(&defaults)
+            end
+
+            __groups.each_value do |group|
+              group.configure_defaults!(environment)
+            end
+          end
+
+          def deprecate(target = self, solution: "do not use")
+            if @__settings.include?(target)
+              super(:"#{target}=", solution: solution)
+            end
+
+            super(target, solution: solution)
+          end
+
+          private def define_setting_methods(name)
+            define_method name do |&block|
+              if block
+                find_setting(name).set(block)
+              else
+                find_setting(name).value
+              end
+            end
+
+            define_method :"#{name}=" do |value|
+              find_setting(name).set(value)
+            end
+          end
+
+          private def define_group_methods(name)
+            define_method name do
+              find_group(name)
+            end
+          end
+
+          private def find_setting(name)
+            @__settings[name]
+          end
+
+          private def find_group(name)
+            @__groups[name]
           end
         end
 
-        # Define defaults for one or more environments.
-        #
-        def defaults(*environments, &block)
-          environments.each do |environment|
-            (@__defaults[environment] ||= Concurrent::Array.new) << block
-          end
+        def initialize(configurable)
+          @configurable = configurable
+          @__settings = Concurrent::Hash.new
+          @__groups = Concurrent::Hash.new
         end
 
-        # Define a nested configurable.
-        #
-        def configurable(group, &block)
-          group = group.to_sym
+        def initialize_copy(_)
+          @__settings = @__settings.deep_dup
+          @__groups = @__groups.deep_dup
 
-          config = Config.new(@configurable, name: group, path: path_to_self)
-          config.instance_eval(&block)
-
-          unless @__groups.include?(group)
-            define_group_methods(group)
-          end
-
-          @__groups[group] = config
-        end
-
-        # Configure default values for `environment`.
-        #
-        def configure_defaults!(environment)
-          @__defaults[environment.to_s.to_sym]&.each do |defaults|
-            instance_eval(&defaults)
-          end
-
-          @__groups.each_pair do |_, group|
-            group.configure_defaults!(environment)
-          end
+          super
         end
 
         # If value for `setting` is a proc, the block will be evaled in `context`. The return value
@@ -125,31 +179,29 @@ module Pakyow
         # Returns configuration as a hash.
         #
         def to_h
+          ensure_state!
+
           hash = {}
 
-          @__settings.each_with_object(hash) { |(name, setting), h|
-            h[name] = setting.value
-          }
+          @__settings.each_pair do |name, setting|
+            hash[name] = setting.value
+          end
 
-          @__groups.each_with_object(hash) { |(name, group), h|
-            h[name] = group.to_h
-          }
+          @__groups.each_pair do |name, group|
+            hash[name] = group.to_h
+          end
 
           hash
         end
 
         def deprecate(target = self, solution: "do not use")
-          define_deprecation(target, solution)
-
-          @__deprecations << { target: target, solution: solution }
+          self.class.deprecate(target, solution: solution)
         end
 
-        private def define_deprecation(target, solution)
-          if @__settings.include?(target)
-            singleton_class.deprecate(:"#{target}=", solution: solution)
-          end
+        def freeze
+          ensure_state!
 
-          singleton_class.deprecate(target, solution: solution)
+          super
         end
 
         # @api private
@@ -163,10 +215,8 @@ module Pakyow
           @__groups.values.each do |group|
             group.update_configurable(configurable)
           end
-        end
 
-        private def path_to_self
-          @__path_to_self ||= (@__path.dup << @__name).compact
+          self
         end
 
         private def deprecated_method_reference(target)
@@ -184,6 +234,10 @@ module Pakyow
               target
             end
 
+            path_to_self = self.class.object_name.parts.reject { |part|
+              part == :config
+            }
+
             @deprecated_method_reference = ([
               configurable_context.name, "config"
             ].concat(path_to_self) << target).compact.join(".")
@@ -193,29 +247,28 @@ module Pakyow
         end
 
         private def find_setting(name)
-          @__settings[name.to_sym]
+          @__settings[name] ||= build_setting(name)
         end
 
         private def find_group(name)
-          @__groups[name.to_sym]
+          @__groups[name] ||= self.class.__groups[name].new(@configurable)
         end
 
-        private def define_setting_methods(name)
-          singleton_class.define_method name do |&block|
-            if block
-              find_setting(name).set(block)
-            else
-              find_setting(name).value
-            end
+        private def build_setting(name)
+          setting = self.class.__settings[name]
+          unless @configurable.equal?(self.class.__configurable)
+            setting = setting.dup.update_configurable(@configurable)
           end
 
-          singleton_class.define_method :"#{name}=" do |value|
-            find_setting(name).set(value)
-          end
+          setting
         end
 
-        private def define_group_methods(name)
-          singleton_class.define_method name do
+        private def ensure_state!
+          self.class.__settings.each_key do |name|
+            find_setting(name)
+          end
+
+          self.class.__groups.each_key do |name|
             find_group(name)
           end
         end