+++ pakyow-support/CHANGELOG.md
@@ -1,5 +1,10 @@
 # v1.1.0 (unreleased)
 
+  * `chg` **Build config classes for configurable objects.**
+
+    *Related links:*
+    - [Pull Request #383][pr-383]
+
   * `chg` **Support passing the isolable context through makeable.**
 
     *Related links:*
@@ -167,6 +172,7 @@
     - [Pull Request #364][pr-364]
 
 [pr-384]: https://github.com/pakyow/pakyow/pull/384
+[pr-383]: https://github.com/pakyow/pakyow/pull/383
 [pr-379]: https://github.com/pakyow/pakyow/pull/379
 [pr-378]: https://github.com/pakyow/pakyow/pull/378
 [pr-377]: https://github.com/pakyow/pakyow/pull/377